convert allocator_tests.cpp
